+++ src/com/steelIndustry/model/MasterCard.java
@@ -25,7 +25,7 @@ public class MasterCard implements BaseSimplePojo {
     @Column(name = "contact")
     private String contact; // 联系人
     @Column(name = "mobile_number")
-    private int mobileNumber; // 联系方式
+    private long mobileNumber; // 联系方式
     @Column(name = "province_id")
     private int provinceId; // 联系地址-省份id
     @Column(name = "province_name")
@@ -91,11 +91,11 @@ public class MasterCard implements BaseSimplePojo {
         this.contact = contact;
     }
 
-    public int getMobileNumber() {
+    public long getMobileNumber() {
         return mobileNumber;
     }
 
-    public void setMobileNumber(int mobileNumber) {
+    public void setMobileNumber(long mobileNumber) {
         this.mobileNumber = mobileNumber;
     }
